requirements.txt
tweepy
praw
requests
pandas
scikit-learn
transformers
statsmodels
schedule
matplotlib
seaborn
flask
python-dotenv
numpy
torch
torchvision
torchaudio